@@ -4,61 +4,61 @@
   "repository": "https://github.com/getfider/fider",
   "private": true,
   "dependencies": {
-    "@lingui/react": "3.10.2",
-    "dompurify": "2.3.1",
-    "marked": "2.1.3",
+    "@lingui/react": "3.12.1",
+    "dompurify": "2.3.3",
+    "marked": "3.0.8",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-textarea-autosize": "8.3.3",
-    "react-toastify": "8.0.2"
+    "react-toastify": "8.0.3"
   },
   "devDependencies": {
-    "@babel/core": "7.15.4",
-    "@babel/preset-env": "7.15.4",
+    "@babel/core": "7.15.8",
+    "@babel/preset-env": "7.15.8",
     "@babel/preset-react": "7.14.5",
     "@babel/preset-typescript": "7.15.0",
     "@cucumber/cucumber": "7.3.1",
-    "@lingui/cli": "3.10.2",
-    "@lingui/loader": "3.10.2",
-    "@lingui/macro": "3.10.2",
+    "@lingui/cli": "3.12.1",
+    "@lingui/loader": "3.12.1",
+    "@lingui/macro": "3.12.1",
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
+    "@testing-library/react": "^12.1.2",
     "@types/debug": "4.1.7",
-    "@types/dompurify": "2.2.3",
+    "@types/dompurify": "2.3.1",
     "@types/jsdom": "16.2.13",
-    "@types/marked": "3.0.0",
-    "@types/node": "16.7.10",
-    "@types/react": "17.0.19",
-    "@types/react-dom": "17.0.9",
+    "@types/marked": "3.0.2",
+    "@types/node": "16.11.6",
+    "@types/react": "17.0.33",
+    "@types/react-dom": "17.0.10",
     "@types/react-textarea-autosize": "4.3.6",
-    "@typescript-eslint/eslint-plugin": "4.30.0",
-    "@typescript-eslint/parser": "4.30.0",
-    "autoprefixer": "10.3.4",
-    "babel-jest": "27.1.0",
-    "babel-loader": "8.2.2",
+    "@typescript-eslint/eslint-plugin": "4.33.0",
+    "@typescript-eslint/parser": "4.33.0",
+    "autoprefixer": "10.4.0",
+    "babel-jest": "27.3.1",
+    "babel-loader": "8.2.3",
     "babel-plugin-macros": "3.1.0",
-    "css-loader": "6.2.0",
-    "esbuild": "0.12.25",
+    "css-loader": "6.5.0",
+    "esbuild": "0.13.10",
     "esbuild-plugin-babel": "0.2.3",
     "eslint": "7.32.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-prettier": "4.0.0",
-    "eslint-plugin-react": "7.25.1",
+    "eslint-plugin-react": "7.26.1",
     "eslint-plugin-react-hooks": "4.2.0",
-    "fork-ts-checker-webpack-plugin": "6.3.3",
+    "fork-ts-checker-webpack-plugin": "6.4.0",
     "isomorphic-fetch": "3.0.0",
-    "jest": "27.1.0",
-    "mini-css-extract-plugin": "2.2.2",
-    "playwright": "1.14.1",
-    "prettier": "2.3.2",
+    "jest": "27.3.1",
+    "mini-css-extract-plugin": "2.4.3",
+    "playwright": "1.16.2",
+    "prettier": "2.4.1",
     "purgecss-webpack-plugin": "4.0.3",
-    "sass": "1.39.0",
-    "sass-loader": "12.1.0",
-    "svg-sprite-loader": "6.0.9",
-    "typescript": "4.4.2",
-    "webpack": "5.52.0",
-    "webpack-bundle-analyzer": "4.4.2",
-    "webpack-cli": "4.8.0"
+    "sass": "1.43.4",
+    "sass-loader": "12.3.0",
+    "svg-sprite-loader": "6.0.11",
+    "typescript": "4.4.4",
+    "webpack": "5.60.0",
+    "webpack-bundle-analyzer": "4.5.0",
+    "webpack-cli": "4.9.1"
   },
   "lingui": {
     "catalogs": [
