@@ -46,7 +46,7 @@ func TestGravatarHandler(t *testing.T) {
 		AddParam("name", user.Name).
 		Execute(handlers.Gravatar())
 
-	expectedAvatar := []byte{255, 216, 255, 224, 0, 16, 74, 70, 73, 70, 0, 1, 1, 1, 0, 96, 0, 96, 0, 0, 255, 254, 0, 59, 67, 82, 69, 65, 84, 79, 82, 58, 32, 103, 100, 45, 106, 112, 101, 103, 32, 118, 49, 46, 48, 32, 40, 117, 115, 105, 110, 103, 32, 73, 74, 71, 32, 74, 80, 69, 71, 32, 118, 54, 50, 41, 44, 32, 113, 117, 97, 108, 105, 116, 121, 32, 61, 32, 57, 48, 10, 255, 219, 0, 67, 0, 3, 2, 2, 3, 2, 2, 3, 3, 3, 3, 4, 3, 3, 4, 5, 8, 5, 5, 4, 4, 5, 10, 7, 7, 6, 8, 12, 10, 12, 12, 11, 10, 11, 11, 13, 14, 18, 16, 13, 14, 17, 14, 11, 11, 16, 22, 16, 17, 19, 20, 21, 21, 21, 12, 15, 23, 24, 22, 20, 24, 18, 20, 21, 20, 255, 219, 0, 67, 1, 3, 4, 4, 5, 4, 5, 9, 5, 5, 9, 20, 13, 11, 13, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 255, 192, 0, 17, 8, 0, 50, 0, 50, 3, 1, 34, 0, 2, 17, 1, 3, 17, 1, 255, 196, 0, 31, 0, 0, 1, 5, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 255, 196, 0, 181, 16, 0, 2, 1, 3, 3, 2, 4, 3, 5, 5, 4, 4, 0, 0, 1, 125, 1, 2, 3, 0, 4, 17, 5, 18, 33, 49, 65, 6, 19, 81, 97, 7, 34, 113, 20, 50, 129, 145, 161, 8, 35, 66, 177, 193, 21, 82, 209, 240, 36, 51, 98, 114, 130, 9, 10, 22, 23, 24, 25, 26, 37, 38, 39, 40, 41, 42, 52, 53, 54, 55, 56, 57, 58, 67, 68, 69, 70, 71, 72, 73, 74, 83, 84, 85, 86, 87, 88, 89, 90, 99, 100, 101, 102, 103, 104, 105, 106, 115, 116, 117, 118, 119, 120, 121, 122, 131, 132, 133, 134, 135, 136, 137, 138, 146, 147, 148, 149, 150, 151, 152, 153, 154, 162, 163, 164, 165, 166, 167, 168, 169, 170, 178, 179, 180, 181, 182, 183, 184, 185, 186, 194, 195, 196, 197, 198, 199, 200, 201, 202, 210, 211, 212, 213, 214, 215, 216, 217, 218, 225, 226, 227, 228, 229, 230, 231, 232, 233, 234, 241, 242, 243, 244, 245, 246, 247, 248, 249, 250, 255, 196, 0, 31, 1, 0, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 255, 196, 0, 181, 17, 0, 2, 1, 2, 4, 4, 3, 4, 7, 5, 4, 4, 0, 1, 2, 119, 0, 1, 2, 3, 17, 4, 5, 33, 49, 6, 18, 65, 81, 7, 97, 113, 19, 34, 50, 129, 8, 20, 66, 145, 161, 177, 193, 9, 35, 51, 82, 240, 21, 98, 114, 209, 10, 22, 36, 52, 225, 37, 241, 23, 24, 25, 26, 38, 39, 40, 41, 42, 53, 54, 55, 56, 57, 58, 67, 68, 69, 70, 71, 72, 73, 74, 83, 84, 85, 86, 87, 88, 89, 90, 99, 100, 101, 102, 103, 104, 105, 106, 115, 116, 117, 118, 119, 120, 121, 122, 130, 131, 132, 133, 134, 135, 136, 137, 138, 146, 147, 148, 149, 150, 151, 152, 153, 154, 162, 163, 164, 165, 166, 167, 168, 169, 170, 178, 179, 180, 181, 182, 183, 184, 185, 186, 194, 195, 196, 197, 198, 199, 200, 201, 202, 210, 211, 212, 213, 214, 215, 216, 217, 218, 226, 227, 228, 229, 230, 231, 232, 233, 234, 242, 243, 244, 245, 246, 247, 248, 249, 250, 255, 218, 0, 12, 3, 1, 0, 2, 17, 3, 17, 0, 63, 0, 252, 207, 109, 22, 21, 199, 12, 61, 112, 213, 57, 209, 237, 33, 182, 105, 165, 46, 168, 163, 183, 57, 62, 149, 245, 7, 236, 5, 225, 255, 0, 3, 235, 159, 181, 46, 155, 99, 227, 221, 31, 79, 214, 244, 137, 244, 107, 150, 182, 179, 213, 33, 89, 160, 123, 133, 101, 32, 148, 111, 148, 144, 130, 67, 200, 237, 94, 239, 255, 0, 5, 55, 248, 123, 240, 166, 203, 194, 254, 23, 210, 126, 29, 120, 111, 194, 158, 24, 214, 165, 214, 86, 43, 171, 171, 8, 161, 179, 45, 7, 145, 35, 54, 66, 224, 20, 12, 83, 45, 142, 14, 209, 223, 7, 179, 151, 70, 212, 79, 38, 51, 189, 175, 61, 207, 139, 79, 193, 22, 210, 244, 49, 168, 222, 70, 100, 114, 155, 150, 44, 0, 188, 142, 6, 113, 147, 214, 188, 207, 84, 125, 62, 210, 249, 173, 96, 178, 44, 85, 182, 231, 130, 73, 252, 171, 236, 77, 122, 230, 211, 83, 248, 123, 13, 181, 133, 237, 189, 237, 205, 148, 40, 151, 83, 7, 85, 141, 36, 85, 220, 9, 98, 113, 180, 237, 206, 115, 192, 53, 243, 111, 135, 188, 51, 164, 71, 227, 107, 121, 117, 45, 95, 79, 186, 182, 50, 3, 32, 176, 147, 206, 216, 51, 215, 177, 35, 220, 10, 43, 69, 69, 46, 84, 42, 18, 114, 187, 155, 216, 169, 160, 248, 58, 195, 92, 133, 198, 207, 42, 120, 240, 94, 41, 19, 4, 3, 208, 253, 56, 171, 147, 124, 54, 180, 83, 247, 98, 3, 25, 232, 223, 227, 95, 74, 254, 206, 158, 16, 240, 47, 141, 255, 0, 110, 47, 15, 120, 71, 80, 180, 77, 127, 194, 122, 134, 149, 112, 37, 242, 103, 150, 5, 44, 45, 222, 85, 108, 198, 202, 195, 13, 26, 140, 103, 191, 61, 13, 125, 161, 241, 7, 246, 40, 248, 19, 166, 207, 226, 11, 75, 123, 61, 86, 11, 230, 137, 26, 22, 143, 85, 148, 165, 129, 59, 137, 3, 113, 57, 227, 105, 195, 238, 60, 142, 153, 171, 167, 21, 39, 203, 107, 179, 58, 142, 80, 138, 159, 53, 147, 63, 36, 207, 195, 155, 44, 253, 216, 255, 0, 241, 239, 254, 42, 138, 250, 238, 251, 224, 231, 192, 219, 11, 219, 139, 89, 126, 37, 220, 44, 144, 200, 209, 178, 181, 237, 190, 65, 7, 4, 31, 147, 218, 138, 233, 246, 17, 236, 190, 244, 115, 123, 106, 157, 223, 226, 121, 191, 236, 225, 225, 237, 38, 235, 196, 90, 30, 175, 127, 4, 114, 125, 140, 93, 100, 191, 241, 6, 27, 54, 253, 48, 77, 124, 243, 241, 167, 198, 35, 226, 39, 196, 237, 94, 235, 78, 139, 26, 104, 185, 107, 125, 58, 218, 17, 144, 33, 86, 33, 118, 129, 221, 190, 241, 247, 106, 244, 107, 47, 23, 73, 225, 15, 133, 90, 133, 196, 13, 229, 220, 75, 20, 144, 68, 71, 4, 51, 182, 220, 143, 112, 9, 63, 133, 120, 5, 178, 204, 242, 55, 144, 28, 176, 82, 78, 204, 228, 14, 245, 203, 94, 126, 226, 130, 245, 59, 240, 112, 213, 212, 126, 136, 244, 13, 55, 74, 178, 240, 239, 135, 124, 77, 160, 235, 186, 69, 235, 248, 154, 254, 218, 6, 210, 86, 221, 247, 164, 18, 137, 163, 102, 50, 42, 57, 249, 140, 94, 114, 224, 130, 65, 113, 144, 59, 114, 87, 90, 123, 216, 216, 219, 93, 164, 166, 11, 168, 159, 99, 194, 210, 1, 42, 54, 73, 4, 40, 59, 128, 192, 28, 144, 57, 53, 235, 31, 7, 143, 143, 126, 30, 106, 54, 94, 32, 240, 236, 210, 232, 154, 157, 131, 75, 50, 203, 106, 26, 43, 150, 67, 20, 76, 81, 164, 94, 118, 178, 49, 194, 156, 242, 132, 17, 212, 87, 97, 241, 150, 255, 0, 196, 30, 54, 240, 54, 163, 172, 248, 150, 120, 237, 158, 105, 197, 202, 202, 2, 192, 215, 239, 130, 0, 242, 219, 27, 74, 156, 238, 0, 99, 43, 129, 201, 192, 227, 183, 83, 177, 212, 229, 146, 91, 220, 227, 127, 103, 175, 137, 175, 225, 175, 142, 190, 1, 241, 13, 228, 134, 40, 99, 186, 58, 117, 212, 209, 49, 70, 141, 103, 86, 133, 159, 35, 160, 11, 38, 239, 248, 11, 12, 98, 189, 215, 246, 171, 241, 135, 142, 45, 245, 173, 91, 77, 209, 124, 79, 126, 44, 245, 139, 159, 54, 243, 124, 196, 201, 147, 144, 64, 110, 170, 164, 99, 229, 30, 149, 241, 45, 148, 139, 230, 52, 82, 62, 200, 228, 227, 118, 126, 235, 127, 11, 126, 127, 161, 53, 245, 126, 187, 227, 120, 254, 34, 120, 99, 70, 241, 35, 184, 150, 226, 242, 216, 45, 216, 207, 41, 118, 152, 89, 129, 244, 203, 124, 224, 127, 118, 68, 174, 202, 18, 188, 92, 14, 76, 92, 57, 92, 106, 37, 177, 226, 171, 240, 158, 2, 163, 204, 190, 99, 39, 241, 125, 123, 209, 91, 242, 106, 110, 36, 108, 103, 25, 52, 86, 156, 144, 254, 83, 139, 219, 213, 254, 98, 75, 109, 16, 120, 175, 192, 113, 105, 198, 65, 19, 191, 207, 28, 135, 160, 96, 73, 25, 246, 237, 248, 215, 157, 234, 191, 12, 124, 83, 160, 193, 113, 52, 154, 116, 173, 104, 138, 124, 201, 237, 220, 58, 21, 235, 147, 131, 211, 128, 121, 244, 174, 183, 71, 241, 141, 151, 135, 60, 63, 100, 247, 83, 22, 148, 198, 25, 97, 143, 150, 111, 240, 252, 107, 152, 241, 151, 197, 61, 91, 197, 182, 198, 199, 119, 216, 244, 204, 130, 109, 227, 57, 47, 142, 155, 219, 191, 211, 165, 101, 87, 217, 181, 119, 185, 215, 134, 85, 163, 39, 20, 189, 219, 255, 0, 86, 56, 239, 51, 228, 219, 142, 125, 107, 214, 62, 16, 252, 25, 241, 47, 199, 75, 235, 75, 107, 101, 187, 155, 78, 179, 30, 76, 183, 140, 75, 45, 180, 99, 36, 12, 177, 194, 168, 207, 78, 131, 53, 15, 134, 252, 1, 168, 63, 129, 228, 241, 11, 91, 88, 106, 0, 198, 124, 184, 46, 98, 102, 109, 163, 142, 10, 144, 115, 199, 124, 213, 95, 17, 106, 30, 37, 248, 117, 161, 157, 30, 211, 84, 159, 79, 178, 212, 80, 61, 205, 189, 155, 180, 105, 38, 224, 172, 59, 231, 161, 0, 250, 226, 185, 121, 26, 179, 107, 67, 181, 212, 231, 188, 96, 245, 48, 190, 38, 248, 13, 62, 29, 120, 170, 243, 71, 77, 98, 203, 89, 16, 57, 95, 58, 206, 64, 227, 232, 112, 72, 4, 122, 100, 214, 175, 195, 127, 19, 182, 157, 165, 221, 233, 247, 26, 149, 188, 86, 79, 47, 157, 29, 172, 163, 14, 36, 192, 5, 149, 136, 192, 4, 0, 8, 207, 59, 87, 142, 5, 112, 166, 202, 115, 100, 110, 202, 17, 0, 144, 71, 188, 247, 98, 9, 199, 228, 42, 189, 17, 151, 44, 174, 141, 37, 14, 120, 114, 201, 158, 181, 37, 234, 51, 177, 15, 193, 39, 24, 52, 87, 147, 110, 62, 166, 138, 232, 250, 199, 145, 195, 245, 37, 252, 223, 128, 249, 216, 179, 140, 146, 112, 160, 115, 244, 162, 220, 3, 113, 16, 35, 32, 176, 227, 241, 162, 138, 228, 61, 37, 178, 62, 222, 240, 253, 180, 48, 120, 14, 241, 34, 137, 35, 65, 230, 0, 168, 160, 1, 197, 121, 119, 199, 187, 120, 158, 227, 74, 221, 26, 55, 250, 29, 185, 229, 71, 247, 18, 138, 43, 210, 173, 252, 63, 184, 249, 236, 63, 241, 126, 103, 156, 120, 158, 24, 227, 248, 96, 74, 162, 169, 254, 215, 65, 144, 49, 199, 144, 213, 230, 180, 81, 94, 124, 186, 30, 229, 47, 132, 40, 162, 138, 147, 99, 255, 217}
+	expectedAvatar := []byte{255, 216, 255, 224, 0, 16, 74, 70, 73, 70, 0, 1, 1, 1, 0, 96, 0, 96, 0, 0, 255, 254, 0, 59, 67, 82, 69, 65, 84, 79, 82, 58, 32, 103, 100, 45, 106, 112, 101, 103, 32, 118, 49, 46, 48, 32, 40, 117, 115, 105, 110, 103, 32, 73, 74, 71, 32, 74, 80, 69, 71, 32, 118, 54, 50, 41, 44, 32, 113, 117, 97, 108, 105, 116, 121, 32, 61, 32, 57, 48, 10, 255, 219, 0, 67, 0, 3, 2, 2, 3, 2, 2, 3, 3, 3, 3, 4, 3, 3, 4, 5, 8, 5, 5, 4, 4, 5, 10, 7, 7, 6, 8, 12, 10, 12, 12, 11, 10, 11, 11, 13, 14, 18, 16, 13, 14, 17, 14, 11, 11, 16, 22, 16, 17, 19, 20, 21, 21, 21, 12, 15, 23, 24, 22, 20, 24, 18, 20, 21, 20, 255, 219, 0, 67, 1, 3, 4, 4, 5, 4, 5, 9, 5, 5, 9, 20, 13, 11, 13, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 255, 192, 0, 17, 8, 0, 50, 0, 50, 3, 1, 34, 0, 2, 17, 1, 3, 17, 1, 255, 196, 0, 31, 0, 0, 1, 5, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 255, 196, 0, 181, 16, 0, 2, 1, 3, 3, 2, 4, 3, 5, 5, 4, 4, 0, 0, 1, 125, 1, 2, 3, 0, 4, 17, 5, 18, 33, 49, 65, 6, 19, 81, 97, 7, 34, 113, 20, 50, 129, 145, 161, 8, 35, 66, 177, 193, 21, 82, 209, 240, 36, 51, 98, 114, 130, 9, 10, 22, 23, 24, 25, 26, 37, 38, 39, 40, 41, 42, 52, 53, 54, 55, 56, 57, 58, 67, 68, 69, 70, 71, 72, 73, 74, 83, 84, 85, 86, 87, 88, 89, 90, 99, 100, 101, 102, 103, 104, 105, 106, 115, 116, 117, 118, 119, 120, 121, 122, 131, 132, 133, 134, 135, 136, 137, 138, 146, 147, 148, 149, 150, 151, 152, 153, 154, 162, 163, 164, 165, 166, 167, 168, 169, 170, 178, 179, 180, 181, 182, 183, 184, 185, 186, 194, 195, 196, 197, 198, 199, 200, 201, 202, 210, 211, 212, 213, 214, 215, 216, 217, 218, 225, 226, 227, 228, 229, 230, 231, 232, 233, 234, 241, 242, 243, 244, 245, 246, 247, 248, 249, 250, 255, 196, 0, 31, 1, 0, 3, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 255, 196, 0, 181, 17, 0, 2, 1, 2, 4, 4, 3, 4, 7, 5, 4, 4, 0, 1, 2, 119, 0, 1, 2, 3, 17, 4, 5, 33, 49, 6, 18, 65, 81, 7, 97, 113, 19, 34, 50, 129, 8, 20, 66, 145, 161, 177, 193, 9, 35, 51, 82, 240, 21, 98, 114, 209, 10, 22, 36, 52, 225, 37, 241, 23, 24, 25, 26, 38, 39, 40, 41, 42, 53, 54, 55, 56, 57, 58, 67, 68, 69, 70, 71, 72, 73, 74, 83, 84, 85, 86, 87, 88, 89, 90, 99, 100, 101, 102, 103, 104, 105, 106, 115, 116, 117, 118, 119, 120, 121, 122, 130, 131, 132, 133, 134, 135, 136, 137, 138, 146, 147, 148, 149, 150, 151, 152, 153, 154, 162, 163, 164, 165, 166, 167, 168, 169, 170, 178, 179, 180, 181, 182, 183, 184, 185, 186, 194, 195, 196, 197, 198, 199, 200, 201, 202, 210, 211, 212, 213, 214, 215, 216, 217, 218, 226, 227, 228, 229, 230, 231, 232, 233, 234, 242, 243, 244, 245, 246, 247, 248, 249, 250, 255, 218, 0, 12, 3, 1, 0, 2, 17, 3, 17, 0, 63, 0, 252, 208, 109, 22, 21, 35, 134, 31, 70, 171, 7, 70, 180, 134, 217, 167, 148, 186, 162, 142, 216, 57, 62, 149, 245, 7, 236, 7, 225, 255, 0, 3, 235, 159, 181, 54, 155, 99, 227, 221, 31, 79, 214, 180, 137, 244, 107, 166, 182, 179, 213, 32, 89, 160, 123, 133, 101, 32, 148, 111, 148, 144, 130, 67, 200, 237, 94, 237, 255, 0, 5, 55, 248, 125, 240, 166, 199, 194, 254, 23, 210, 126, 29, 120, 111, 194, 158, 24, 214, 165, 214, 86, 43, 171, 171, 8, 161, 179, 45, 7, 145, 35, 54, 66, 224, 20, 12, 83, 45, 142, 14, 209, 223, 158, 181, 23, 102, 212, 79, 38, 50, 189, 175, 61, 207, 139, 79, 193, 22, 211, 52, 49, 168, 222, 70, 100, 114, 155, 150, 44, 0, 188, 142, 6, 113, 147, 212, 87, 153, 234, 146, 105, 246, 183, 205, 107, 13, 150, 226, 173, 183, 60, 18, 79, 229, 95, 98, 107, 215, 22, 154, 159, 195, 216, 109, 172, 47, 109, 239, 110, 108, 161, 141, 46, 166, 14, 171, 26, 72, 171, 144, 75, 19, 141, 167, 110, 115, 158, 1, 175, 155, 188, 61, 225, 157, 34, 63, 27, 91, 203, 169, 106, 250, 125, 213, 177, 144, 52, 130, 194, 79, 63, 96, 207, 94, 196, 143, 112, 42, 171, 69, 69, 46, 84, 20, 36, 229, 204, 230, 246, 41, 232, 62, 13, 176, 215, 33, 127, 147, 202, 184, 143, 5, 226, 145, 48, 64, 61, 8, 246, 226, 174, 205, 240, 222, 209, 115, 242, 196, 6, 51, 209, 191, 198, 190, 148, 253, 157, 60, 33, 224, 79, 27, 254, 220, 94, 30, 240, 142, 161, 104, 186, 255, 0, 132, 245, 13, 42, 224, 75, 228, 207, 44, 11, 184, 91, 188, 170, 217, 141, 149, 134, 26, 53, 227, 61, 249, 232, 107, 237, 31, 136, 63, 177, 79, 192, 157, 58, 127, 16, 90, 91, 217, 234, 176, 95, 52, 72, 208, 152, 245, 89, 10, 88, 22, 220, 78, 55, 19, 158, 54, 156, 62, 227, 200, 233, 154, 170, 113, 83, 124, 188, 183, 102, 85, 28, 161, 21, 62, 107, 38, 126, 73, 31, 135, 54, 89, 60, 71, 255, 0, 143, 255, 0, 241, 84, 87, 215, 151, 223, 7, 62, 7, 88, 94, 220, 90, 203, 241, 42, 225, 100, 130, 70, 141, 149, 175, 109, 242, 8, 56, 32, 252, 158, 212, 87, 79, 176, 143, 101, 247, 163, 155, 219, 84, 238, 255, 0, 19, 205, 191, 103, 15, 15, 105, 55, 94, 34, 208, 245, 123, 251, 120, 228, 54, 98, 235, 37, 191, 136, 48, 217, 183, 233, 130, 127, 58, 249, 231, 227, 79, 140, 135, 196, 79, 137, 218, 189, 214, 157, 30, 52, 209, 114, 214, 250, 117, 180, 35, 32, 66, 172, 66, 144, 7, 118, 251, 199, 221, 171, 209, 172, 188, 93, 39, 132, 62, 21, 106, 23, 16, 55, 151, 113, 44, 82, 65, 17, 28, 16, 206, 197, 114, 61, 192, 36, 254, 21, 224, 22, 171, 59, 200, 222, 70, 242, 193, 73, 98, 157, 64, 239, 92, 181, 231, 238, 40, 47, 83, 191, 7, 13, 93, 71, 232, 143, 65, 211, 116, 171, 47, 14, 248, 119, 196, 218, 6, 187, 164, 94, 191, 137, 175, 237, 160, 125, 36, 91, 190, 244, 183, 148, 77, 27, 49, 145, 81, 207, 204, 98, 243, 151, 4, 18, 11, 12, 129, 219, 145, 186, 176, 146, 198, 198, 218, 237, 36, 48, 221, 70, 251, 30, 22, 148, 9, 81, 178, 72, 33, 65, 220, 6, 0, 228, 129, 201, 175, 88, 248, 60, 124, 123, 240, 239, 81, 178, 241, 7, 135, 102, 151, 68, 213, 44, 90, 89, 146, 107, 80, 209, 92, 148, 49, 68, 197, 26, 69, 231, 107, 35, 18, 20, 231, 230, 66, 8, 234, 43, 176, 248, 203, 127, 175, 248, 219, 192, 218, 142, 179, 226, 89, 227, 182, 146, 105, 197, 200, 148, 5, 129, 175, 223, 4, 1, 229, 182, 54, 149, 57, 220, 0, 198, 87, 3, 147, 129, 199, 110, 167, 99, 169, 202, 237, 189, 206, 55, 246, 122, 248, 156, 254, 26, 248, 235, 224, 31, 16, 222, 72, 98, 134, 59, 163, 167, 93, 77, 19, 20, 104, 214, 117, 104, 89, 242, 58, 0, 178, 110, 255, 0, 128, 176, 198, 43, 221, 127, 106, 191, 24, 120, 226, 223, 90, 213, 180, 221, 23, 196, 215, 226, 207, 88, 185, 243, 111, 55, 204, 76, 153, 57, 4, 7, 234, 170, 70, 62, 81, 233, 95, 18, 216, 200, 166, 70, 138, 71, 217, 28, 188, 111, 207, 221, 111, 225, 111, 207, 244, 38, 190, 175, 215, 124, 111, 31, 196, 79, 12, 104, 222, 36, 119, 18, 220, 221, 219, 5, 187, 25, 229, 46, 211, 11, 48, 62, 153, 111, 156, 15, 238, 200, 149, 217, 66, 87, 139, 129, 201, 139, 135, 43, 141, 69, 208, 241, 101, 248, 79, 1, 81, 230, 95, 49, 147, 31, 55, 29, 251, 209, 91, 242, 106, 114, 9, 27, 25, 198, 79, 122, 43, 79, 103, 15, 229, 56, 189, 189, 95, 230, 99, 237, 180, 65, 226, 207, 1, 197, 167, 25, 68, 82, 63, 207, 28, 135, 160, 96, 73, 25, 246, 237, 248, 215, 157, 234, 191, 12, 60, 83, 160, 193, 113, 52, 154, 116, 166, 209, 20, 249, 147, 219, 184, 116, 42, 57, 201, 193, 233, 192, 60, 250, 87, 93, 163, 248, 198, 203, 195, 158, 31, 178, 123, 169, 137, 148, 198, 25, 97, 143, 150, 111, 240, 252, 107, 151, 241, 151, 197, 77, 91, 197, 214, 198, 199, 119, 216, 244, 204, 228, 219, 198, 114, 100, 199, 77, 237, 223, 233, 210, 178, 171, 236, 218, 187, 220, 235, 195, 42, 209, 147, 138, 94, 237, 250, 254, 135, 29, 230, 29, 155, 113, 207, 173, 122, 207, 194, 31, 131, 62, 37, 248, 233, 125, 105, 109, 110, 183, 115, 233, 214, 99, 201, 150, 241, 137, 116, 182, 140, 100, 129, 150, 56, 85, 30, 157, 6, 106, 15, 13, 248, 3, 80, 127, 3, 201, 226, 22, 182, 176, 212, 65, 140, 249, 112, 92, 196, 204, 219, 71, 28, 21, 32, 231, 142, 249, 170, 222, 34, 212, 60, 75, 240, 235, 67, 58, 61, 166, 169, 62, 159, 101, 168, 160, 123, 155, 123, 55, 104, 210, 77, 193, 88, 119, 207, 66, 1, 245, 197, 114, 242, 53, 103, 37, 161, 218, 234, 41, 222, 48, 122, 152, 31, 19, 124, 4, 159, 14, 188, 85, 121, 163, 166, 177, 101, 172, 136, 28, 175, 157, 103, 32, 113, 244, 56, 36, 2, 61, 50, 107, 91, 225, 183, 137, 219, 78, 210, 238, 244, 251, 141, 74, 222, 43, 39, 151, 206, 142, 214, 81, 135, 18, 96, 2, 202, 196, 96, 2, 48, 8, 207, 59, 87, 142, 5, 112, 173, 101, 112, 108, 77, 217, 140, 249, 2, 65, 30, 243, 221, 136, 39, 31, 144, 254, 85, 90, 136, 203, 150, 92, 200, 185, 67, 218, 67, 146, 76, 245, 201, 47, 148, 187, 16, 220, 18, 113, 131, 69, 121, 38, 226, 59, 209, 93, 31, 88, 242, 56, 190, 164, 191, 155, 240, 31, 112, 197, 156, 100, 147, 133, 3, 159, 165, 22, 192, 27, 136, 129, 25, 5, 199, 7, 235, 69, 21, 200, 122, 75, 100, 125, 193, 225, 251, 104, 96, 240, 21, 218, 69, 18, 70, 131, 204, 1, 81, 64, 3, 138, 242, 223, 143, 86, 241, 61, 198, 149, 186, 36, 111, 244, 59, 115, 202, 143, 238, 37, 20, 87, 167, 91, 248, 127, 113, 243, 244, 63, 137, 243, 60, 227, 196, 240, 199, 31, 194, 242, 85, 21, 79, 246, 186, 12, 129, 142, 60, 134, 175, 52, 162, 138, 243, 165, 208, 246, 168, 252, 33, 69, 20, 84, 155, 159, 255, 217}
 	Expect(code).Equals(http.StatusOK)
 	Expect(response.Body.Bytes()).Equals(expectedAvatar)
 }
