@@ -2,7 +2,9 @@ package jobs
 
 import (
 	"github.com/getfider/fider/app/models/cmd"
+	"github.com/getfider/fider/app/models/dto"
 	"github.com/getfider/fider/app/pkg/bus"
+	"github.com/getfider/fider/app/pkg/log"
 )
 
 type PurgeExpiredNotificationsJobHandler struct {
@@ -13,5 +15,17 @@ func (e PurgeExpiredNotificationsJobHandler) Schedule() string {
 }
 
 func (e PurgeExpiredNotificationsJobHandler) Run(ctx Context) error {
-	return bus.Dispatch(ctx, &cmd.PurgeExpiredNotifications{})
+	log.Debug(ctx, "deleting notifications older than 1 year")
+
+	c := &cmd.PurgeExpiredNotifications{}
+	err := bus.Dispatch(ctx, c)
+	if err != nil {
+		return err
+	}
+
+	log.Debugf(ctx, "@{RowsDeleted} notifications were deleted", dto.Props{
+		"RowsDeleted": c.NumOfDeletedNotifications,
+	})
+
+	return nil
 }
