@@ -3,6 +3,7 @@ package middlewares
 import (
 	"net/http"
 
+	"github.com/getfider/fider/app/models/dto"
 	"github.com/getfider/fider/app/models/enum"
 	"github.com/getfider/fider/app/models/query"
 
@@ -10,6 +11,7 @@ import (
 	"github.com/getfider/fider/app/pkg/bus"
 	"github.com/getfider/fider/app/pkg/env"
 	"github.com/getfider/fider/app/pkg/errors"
+	"github.com/getfider/fider/app/pkg/log"
 	"github.com/getfider/fider/app/pkg/web"
 )
 
@@ -31,10 +33,14 @@ func SingleTenant() web.MiddlewareFunc {
 				return c.Failure(err)
 			}
 
-			if firstTenant.Result != nil && firstTenant.Result.Status != enum.TenantDisabled {
+			if firstTenant.Result != nil && !firstTenant.Result.IsDisabled() {
 				c.SetTenant(firstTenant.Result)
 
 				if c.Request.URL.Hostname() != env.Config.HostDomain {
+					log.Errorf(c, "Requested hostname '@{URLHostname}' does not match environment HOST_DOMAIN '@{HostDomain}'.", dto.Props{
+						"URLHostname": c.Request.URL.Hostname(),
+						"HostDomain":  env.Config.HostDomain,
+					})
 					return c.NotFound()
 				}
 			}
@@ -65,7 +71,7 @@ func MultiTenant() web.MiddlewareFunc {
 				return c.Failure(err)
 			}
 
-			if byDomain.Result != nil && byDomain.Result.Status != enum.TenantDisabled {
+			if byDomain.Result != nil && !byDomain.Result.IsDisabled() {
 				c.SetTenant(byDomain.Result)
 
 				if byDomain.Result.CNAME != "" && !c.IsAjax() {
@@ -126,19 +132,14 @@ func CheckTenantPrivacy() web.MiddlewareFunc {
 	}
 }
 
-// BlockLockedTenants blocks requests of non-administrator users on locked tenants
+// BlockLockedTenants blocks requests on locked tenants as they are in read-only mode
 func BlockLockedTenants() web.MiddlewareFunc {
 	return func(next web.HandlerFunc) web.HandlerFunc {
 		return func(c *web.Context) error {
 			if c.Tenant().Status == enum.TenantLocked {
-				if c.Request.IsAPI() {
-					return c.JSON(http.StatusLocked, web.Map{})
-				}
 
-				isAdmin := c.IsAuthenticated() && c.User().Role == enum.RoleAdministrator
-				if !isAdmin {
-					return c.Redirect("/signin")
-				}
+				// Only API operations are blocked, so it's ok to always return a JSON
+				return c.JSON(http.StatusPaymentRequired, web.Map{})
 			}
 			return next(c)
 		}
