@@ -2,6 +2,7 @@ package jwt
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/getfider/fider/app/pkg/env"
 	"github.com/getfider/fider/app/pkg/errors"
@@ -11,7 +12,11 @@ import (
 var jwtSecret = env.Config.JWTSecret
 
 // Metadata is the basic JWT information
-type Metadata = jwtgo.StandardClaims
+type Metadata = jwtgo.RegisteredClaims
+
+func Time(t time.Time) *jwtgo.NumericDate {
+	return jwtgo.NewNumericDate(t)
+}
 
 const (
 	//FiderClaimsOriginUI is assigned to Fider claims when the Auth Token is generated through the UI
