@@ -65,7 +65,7 @@ func TestJWT_DecodeExpired(t *testing.T) {
 		UserName:  "Jon Snow",
 		UserEmail: "jon.snow@got.com",
 		Metadata: jwt.Metadata{
-			ExpiresAt: time.Now().Unix(),
+			ExpiresAt: jwt.Time(time.Now()),
 		},
 	}
 
@@ -123,7 +123,7 @@ func TestJWT_DecodeOAuthClaimsExpired(t *testing.T) {
 		OAuthName:     "Jon Snow",
 		OAuthProvider: "facebook",
 		Metadata: jwt.Metadata{
-			ExpiresAt: time.Now().Unix(),
+			ExpiresAt: jwt.Time(time.Now()),
 		},
 	}
 
