@@ -38,7 +38,12 @@ func (u *dbUser) toModel(ctx context.Context) *entity.User {
 		return nil
 	}
 
+	avatarURL := ""
 	avatarType := enum.AvatarType(u.AvatarType.Int64)
+	if u.AvatarType.Valid {
+		avatarURL = buildAvatarURL(ctx, avatarType, int(u.ID.Int64), u.Name.String, u.AvatarBlobKey.String)
+	}
+
 	user := &entity.User{
 		ID:            int(u.ID.Int64),
 		Name:          u.Name.String,
@@ -49,7 +54,7 @@ func (u *dbUser) toModel(ctx context.Context) *entity.User {
 		Status:        enum.UserStatus(u.Status.Int64),
 		AvatarType:    avatarType,
 		AvatarBlobKey: u.AvatarBlobKey.String,
-		AvatarURL:     buildAvatarURL(ctx, avatarType, int(u.ID.Int64), u.Name.String, u.AvatarBlobKey.String),
+		AvatarURL:     avatarURL,
 	}
 
 	for i, p := range u.Providers {
