@@ -79,7 +79,9 @@ func TestNotifyAboutDeletePostTask(t *testing.T) {
 		"change":   "<a href='http://domain.com/settings'>change your notification preferences</a>",
 		"logo":     "https://getfider.com/images/logo-100x100.png",
 	})
-	Expect(emailmock.MessageHistory[0].From).Equals("Jon Snow")
+	Expect(emailmock.MessageHistory[0].From).Equals(dto.Recipient{
+		Name: "Jon Snow",
+	})
 	Expect(emailmock.MessageHistory[0].To).HasLen(1)
 	Expect(emailmock.MessageHistory[0].To[0]).Equals(dto.Recipient{
 		Name:    "Arya Stark",
