@@ -75,7 +75,9 @@ func TestNotifyAboutNewCommentTask(t *testing.T) {
 		"unsubscribe": "<a href='http://domain.com/posts/1/add-support-for-typescript'>unsubscribe from it</a>",
 		"logo":        "https://getfider.com/images/logo-100x100.png",
 	})
-	Expect(emailmock.MessageHistory[0].From).Equals("Arya Stark")
+	Expect(emailmock.MessageHistory[0].From).Equals(dto.Recipient{
+		Name: "Arya Stark",
+	})
 	Expect(emailmock.MessageHistory[0].To).HasLen(1)
 	Expect(emailmock.MessageHistory[0].To[0]).Equals(dto.Recipient{
 		Name:    "Jon Snow",
