@@ -32,7 +32,9 @@ func TestSendSignUpEmailTask(t *testing.T) {
 	Expect(emailmock.MessageHistory[0].Props).Equals(dto.Props{
 		"logo": "https://getfider.com/images/logo-100x100.png",
 	})
-	Expect(emailmock.MessageHistory[0].From).Equals("Fider")
+	Expect(emailmock.MessageHistory[0].From).Equals(dto.Recipient{
+		Name: "Fider",
+	})
 	Expect(emailmock.MessageHistory[0].To).HasLen(1)
 	Expect(emailmock.MessageHistory[0].To[0]).Equals(dto.Recipient{
 		Props: dto.Props{
