@@ -39,7 +39,11 @@ describe("<VoteCounter />", () => {
     post.hasVoted = true
     post.votesCount = 9
 
-    const { container } = render(<VoteCounter post={post} />)
+    const { container } = render(
+      <FiderContext.Provider value={fiderMock.authenticated()}>
+        <VoteCounter post={post} />
+      </FiderContext.Provider>
+    )
     const button = container.querySelector("button")
 
     expect(button).toHaveTextContent("9")
@@ -50,7 +54,11 @@ describe("<VoteCounter />", () => {
   test("when hasVoted === false", () => {
     post.hasVoted = false
     post.votesCount = 2
-    const { container } = render(<VoteCounter post={post} />)
+    const { container } = render(
+      <FiderContext.Provider value={fiderMock.authenticated()}>
+        <VoteCounter post={post} />
+      </FiderContext.Provider>
+    )
     const button = container.querySelector("button")
     expect(button).toHaveTextContent("2")
     expect(button).not.toHaveClass("c-vote-counter__button--voted")
@@ -59,7 +67,11 @@ describe("<VoteCounter />", () => {
 
   test("when post is closed", () => {
     post.status = PostStatus.Completed.value
-    const { container } = render(<VoteCounter post={post} />)
+    const { container } = render(
+      <FiderContext.Provider value={fiderMock.authenticated()}>
+        <VoteCounter post={post} />
+      </FiderContext.Provider>
+    )
     const button = container.querySelector("button")
     expect(button).toHaveTextContent("5")
     expect(button).toHaveClass("c-vote-counter__button--disabled")
@@ -83,8 +95,6 @@ describe("<VoteCounter />", () => {
   })
 
   test("click when authenticated and hasVoted === false", async () => {
-    fiderMock.authenticated()
-
     const mock = httpMock.alwaysOk()
 
     const { container } = render(
@@ -106,7 +116,6 @@ describe("<VoteCounter />", () => {
 
   test("click when authenticated and hasVoted === true", async () => {
     post.hasVoted = true
-    fiderMock.authenticated()
 
     const mock = httpMock.alwaysOk()
 
