@@ -7,7 +7,7 @@ interface StackProps {
   onClick?: () => void
   center?: boolean
   divide?: boolean
-  justify?: "between" | "evenly" | "full"
+  justify?: "between" | "evenly" | "full" | "center"
   spacing?: 0 | 1 | 2 | 4 | 6 | 8
 }
 
@@ -24,6 +24,7 @@ const Stack = (props: StackProps, dir: "x" | "y") => {
     "justify-between": props.justify === "between",
     "justify-evenly": props.justify === "evenly",
     "justify-full": props.justify === "full",
+    "justify-center": props.justify === "center",
   })
 
   return (
